[Windows] Fix Visual Studio warnings (we have quite a lot of them)

+++ xbmc/cores/ExternalPlayer/ExternalPlayer.cpp
@@ -592,7 +592,7 @@ void CExternalPlayer::SetSpeed(float iSpeed)
 
 float CExternalPlayer::GetSpeed()
 {
-  return m_speed;
+  return (float)m_speed;
 }
 
 void CExternalPlayer::ShowOSD(bool bOnoff)
